refactor: add models e repositores
